Drink Kool-Aid.
The previous approach of subverting the library's doubly-linked lists
and fixing pointers when deemed necessary was causing all kinds of crazy
problems, because you're not supposed to do that.  Since that was all an
insane micro-optimization anyway (cloning an image just copies a few
numbers, not the actual image data), I've rewritten basically everything
to use plain old image list manipulation instead.  And now it doesn't
crash with an assertion failure.  Hooray.