Write out cffi configuration to a JSON file.
This has several advantages. First, pkg-config won't have to be called on every
sanpera import. Second, built binaries will be more portable across systems.
cffi uses the CFLAGS/LDFLAGS as part of the input into its hashed module names,
so if a binary is shipped to a machine that doesn't have knowledge of what the
CFLAGS and LDFLAGS were, it won't have to guess (and therefore not find the
module).